✨ feat(book/sentence): simplify sentence screen to display book information
- remove unused code related to camera functionality and text recognition
- add new components to display book sentence or a guide if no book selected
- update text in book-frame-guide.tsx
- add new files: book-sentence.tsx and book-sentence-guide.tsx to handle book sentence display
(Note: The creation of new file sentence.tsx was omitted as it appears to be a copy of the file from which content was removed.)
