+++ apps/tk-capture-book/features/book/components/setence/book-sentence.tsx
@@ -2,10 +2,10 @@ import { View } from "react-native";
 
 import { Text } from "@/components/ui/text";
 
-export function BookFrameGuide() {
+export function BookSentence() {
   return (
     <View className="flex-1 items-center justify-center">
-      <Text>BookFrameGuide</Text>
+      <Text>BookSentence</Text>
     </View>
   );
 }
